rename /surveys endpoint to /survey/list
