+++ app/src/features/surveys/view/SurveyHeader.tsx
@@ -124,7 +124,7 @@ const SurveyHeader = () => {
         return;
       }
 
-      history.push(`/admin/projects/${surveyContext.projectId}/surveys`);
+      history.push(`/admin/projects/${surveyContext.projectId}/survey/lits`);
     } catch (error) {
       const apiError = error as APIError;
       showDeleteErrorDialog({ dialogText: apiError.message, open: true });
